public/releases/bluefusion-v1.0.0/requirements.txt
# BlueFusion Requirements
# Auto-generated from pyproject.toml

bleak>=0.21.0
pyserial>=3.5
asyncio>=3.4.3
pydantic>=2.0
rich>=13.0
numpy>=1.24.0
pandas>=2.0.0
aiofiles>=23.0
structlog>=23.0
click>=8.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
gradio>=4.0.0
httpx>=0.25.0
plotly>=5.17.0